Generate a number from a numeric vector. Uses digits_from() output to generate the numeric value of the text.

Generate a number from a numeric vector. Uses digits_from() output to generate the numeric value of the text.

Usage

number_from(digits)

Arguments

digits

A numeric vector translated from words.

Value

A numeric value.